Separate all command classes
